Shoppingcart 'fixed'

+++ src/main/java/com/codecool/shop/dao/implementation/JDBC/ShoppingCartDaoJDBC.java
@@ -5,6 +5,7 @@ import com.codecool.shop.dao.ProductDao;
 import com.codecool.shop.dao.ShoppingCartDao;
 import com.codecool.shop.model.Product;
 
+import javax.servlet.http.HttpSession;
 import java.sql.*;
 import java.util.*;
 
@@ -21,21 +22,24 @@ public class ShoppingCartDaoJDBC implements ShoppingCartDao {
     public static String phone_number =DEFAULT_STRING;
     public static String email_address =DEFAULT_STRING;
     private int orderNumber;
-    Random random = new Random();
+    private int userId;
+    private Random random = new Random();
 
     private int checkNumber;
 
-    public ShoppingCartDaoJDBC(){
+    public ShoppingCartDaoJDBC(int userId){
         this.orderNumber = random.nextInt(10000);
+        this.userId = userId;
     }
 
     @Override
     public void add(Product product) {
         try(Connection con = getConnection()) {
-            String query = "INSERT INTO shoppingcart(user_id, product_id, order_number) VALUES (1,?,?);";
+            String query = "INSERT INTO shoppingcart(user_id, product_id, order_number) VALUES (?,?,?);";
             PreparedStatement statement = con.prepareStatement(query);
-            statement.setInt(1, product.getId());
-            statement.setInt(2, orderNumber);
+            statement.setInt(1, userId);
+            statement.setInt(2, product.getId());
+            statement.setInt(3, orderNumber);
             statement.execute();
             ++checkNumber;
 
@@ -47,10 +51,11 @@ public class ShoppingCartDaoJDBC implements ShoppingCartDao {
     @Override
     public void remove(int id) {
         try(Connection con = getConnection()) {
-            String query = "DELETE FROM shoppingcart WHERE id = (SELECT id FROM shoppingcart WHERE product_id = ? AND order_number = ? LIMIT 1);";
+            String query = "DELETE FROM shoppingcart WHERE id = (SELECT id FROM shoppingcart WHERE product_id = ? AND order_number = ? LIMIT 1) AND user_id = ?;";
             PreparedStatement statement = con.prepareStatement(query);
             statement.setInt(1, id);
             statement.setInt(2, orderNumber);
+            statement.setInt(3, userId);
             statement.execute();
             --checkNumber;
 
@@ -65,9 +70,10 @@ public class ShoppingCartDaoJDBC implements ShoppingCartDao {
         Map<Product, Integer> shoppingCartMap = new HashMap<Product, Integer>();
 
         try(Connection con = getConnection()) {
-            String query = "SELECT product_id FROM shoppingcart WHERE order_number = ? AND 'user' = 'user';";
+            String query = "SELECT product_id FROM shoppingcart WHERE order_number = ? AND user_id = ?;";
             PreparedStatement statement = con.prepareStatement(query);
             statement.setInt(1, orderNumber);
+            statement.setInt(2, userId);
             ResultSet results = statement.executeQuery();
             while (results.next()){
                 Product currentProduct = productDao.find(results.getInt("product_id"));
@@ -94,9 +100,10 @@ public class ShoppingCartDaoJDBC implements ShoppingCartDao {
     public void clear() {
         checkNumber = 0;
         try(Connection con = getConnection()) {
-            String query = "DELETE FROM shoppingcart WHERE order_number = ? AND 'user' = 'user';";
+            String query = "DELETE FROM shoppingcart WHERE order_number = ? AND user_id = ?;";
             PreparedStatement statement = con.prepareStatement(query);
             statement.setInt(1, orderNumber);
+            statement.setInt(2, userId);
             statement.execute();
 
         } catch (SQLException ex) {
